requirements.txt
Flask==2.3.2
Flask-MySQLdb==1.0.1
mysqlclient==2.2.0
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7
pandas==2.1.1
openpyxl==3.1.2 